phase1_stream_to_frame/requirements.txt
streamlink>=5
opencv-python-headless>=4
m3u8
kafka-python>=2.0.1
numpy
